Change the project structure
Since serialization is implemented in two .hpp files, I decided to completely remove the serialize_deserialize subproject. The header files are now located in tests and usage_example.
I also implemented the ability to select the desired implementation (sfinae or concepts) using the USE_CONCEPTS constant, which is defined in CMake and checked when included in the file.